shadow-python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.26.0
python-dotenv>=1.0.0
supabase>=2.0.0
openai>=1.12.0
deepgram-sdk>=3.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
python-dateutil>=2.8.2
structlog>=24.1.0
aiofiles>=23.2.1
pytest>=8.3.3
pytest-asyncio>=0.23.7
pytest-mock>=3.14.0
pytest-cov>=5.0.0
PyJWT>=2.9.0

